+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2/2/25
-- added a default texture to GLT
+- added a default texture to GLT (accessible by passing `GLR_DEFAULT_TEXTURE` instead of a proper path)
 - fixed renderer to support more than 1 entity at a time
+- fixed depth test
 
 ### 1/2/25
 - created `Vec4` (`utils/vec4.h`)